refactor: rename Button component and update modal implementations for workspace management
